fix: fix not trigger GC before take snapshot
